fix(metrics): restructure feature with proper Nx architecture and fix import paths
- Create proper Nx project structure with project.json and tsconfig files
- Organize code into clean layers: application, infrastructure, presentation
- Fix broken imports from non-existent libs paths
- Consolidate exports into single index.ts file
- Add path alias @portals/shared/features/metrics
- Update all references to use new structure
- Fix TypeScript compilation errors and type issues
