+++ apps/portals/shared/features/metrics/project.json
@@ -1,0 +1,38 @@
+{
+  "name": "metrics",
+  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "prefix": "metrics",
+  "sourceRoot": "apps/portals/shared/features/metrics/src",
+  "tags": ["type:metrics", "scope:statistics"],
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/apps/portals/shared/features/metrics",
+        "main": "apps/portals/shared/features/metrics/src/index.ts",
+        "tsConfig": "apps/portals/shared/features/metrics/tsconfig.lib.json",
+        "assets": ["apps/portals/shared/features/metrics/*.md"]
+      }
+    },
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "apps/portals/shared/features/metrics/jest.config.ts",
+        "passWithNoTests": true
+      },
+      "configurations": {
+        "ci": {
+          "ci": true,
+          "codeCoverage": true
+        }
+      }
+    },
+    "lint": {
+      "executor": "@nx/eslint:lint",
+      "outputs": ["{options.outputFile}"]
+    }
+  }
+}
