+++ apps/portals/shared/features/metrics/src/application/models/estimated-user-span.dto.ts
@@ -1,0 +1,4 @@
+export interface EstimatedUserSpanDto {
+  id: number;
+  name: string;
+}
