+++ apps/portals/shared/features/metrics/src/index.ts
@@ -1,0 +1,13 @@
+// Application layer exports
+export * from './application/user-statistic.service';
+export * from './application/ports/estimated-user-span-provider.port';
+export * from './application/models/estimated-user-span.dto';
+
+// Infrastructure layer exports
+export * from './infrastructure/estimated-user-span-api.service';
+
+// Presentation layer exports
+export * from './presentation/estimated-user-span-list-container.directive';
+
+// Feature providers
+export * from './user-statistic.providers';
