+++ apps/portals/shared/features/metrics/src/infrastructure/estimated-user-span-api.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@angular/core";
 import { Observable, of } from "rxjs";
-import { IEstimatedUserSpanProvider, EstimatedUserSpanDto } from "../../../../../../../libs/features/listing/statistic/users/application";
+import { IEstimatedUserSpanProvider } from "../application/ports/estimated-user-span-provider.port";
+import { EstimatedUserSpanDto } from "../application/models/estimated-user-span.dto";
 import { Result } from "@standard";
 
 @Injectable()
@@ -8,6 +9,7 @@ export class EstimatedUserSpanApiService implements IEstimatedUserSpanProvider {
 
   public getEstimatedUserSpans(): Observable<Result<EstimatedUserSpanDto[], Error>> {
     return of({
+      ok: true,
       value: [
         { id: 0, name: "0-1000" },
         { id: 1, name: "1000-10000" },
