+++ apps/portals/shared/features/metrics/src/user-statistic.providers.ts
@@ -1,5 +1,6 @@
 import { ApplicationConfig } from "@angular/core";
-import { ESTIMATED_USER_SPAN_PROVIDER, UserStatisticService } from "../../../../../../../libs/features/listing/statistic/users/application";
+import { ESTIMATED_USER_SPAN_PROVIDER } from "./application/ports/estimated-user-span-provider.port";
+import { UserStatisticService } from "./application/user-statistic.service";
 import { EstimatedUserSpanApiService } from "./infrastructure/estimated-user-span-api.service";
 
 export function provideUserStatisticPlatformFeature(): ApplicationConfig {
