+++ prompts.md
@@ -7,28 +7,27 @@ steps:
 
 1. workon:
 
-2. switch to main
+2. create branch based on main related to de topic -> workon:
 
-3. create branch related to de topic -> workon:
-
-4. Make actual operations on workon
+3. Make actual operations on workon
 fix imports in this feature. Make changes only in related files. Do not move files. Do not change logic in files. You can change only configuration code/naming. Do not rename files. Fix nx boilerplate if needed. Add nx boilerplate if needed. Stick to the architectural conventions (First learn them). Add aliases by root tsconfig. Use tsconfigs from identity. Scan lib folder because thera are most of missing imports. 
 
-5. add conventional commit, push to origin, connect to github, create merge request, accept merge request.
+4. add feature branch based on main, add conventional commit, push to origin, connect to github cli, create merge request, accept merge request.
 
-6. Remove local branch.
-
-
+5. Remove local branch.
 
 
 
 
 
 
+add feature branch based on main, add conventional commit, push to origin, connect to github by cli, create merge request, accept merge request.
 
 
 
 
 
 
-add feature branch, add conventional commit, push to origin, connect to github, create merge request, accept merge request.
+
+
+
