+++ tsconfig.base.json
@@ -68,7 +68,8 @@
       "@domains/catalog/entry": ["libs/domains/catalog/entry/application/index.ts"],
       "@domains/catalog/compatibility": ["libs/domains/catalog/compatibility/application/index.ts"],
       "@domains/identity": ["libs/domains/identity/index.ts"],
-      "@portals/shared/features/filtering": ["apps/portals/shared/features/filtering/src/index.ts"]
+      "@portals/shared/features/filtering": ["apps/portals/shared/features/filtering/src/index.ts"],
+      "@portals/shared/features/metrics": ["apps/portals/shared/features/metrics/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"],
